# perf(node): use faster utf8 byte length in Buffer#from